requirements.txt
google-api-core==1.12.0
google-api-python-client==1.7.11 
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
oauth2client==4.1.3

wheel 
uvicorn 
fastapi 
pydantic
databases
pytest
aiohttp
ujson
python-dotenv
alembic

tortoise-orm
aiomysql
asyncpg
aiosqlite